* For better debugging, the project has moved to MSVC/VS instead of MinGW/CB. Building through the old way is still possible. * C++ compatibility changed to C++17. * Redone batching architecture, separated center/corners and batching is now done on separate thread for better performance. * A client-side console was added for issuing general-purpose commands and ease testing of new features. * Many changes to the GUI/Component system.

+++ src/Render/Texture.cpp
@@ -38,7 +38,7 @@ void Texture::set(const Image& image){
 	height = image.height;
 
 	baseTexture.reset(new uint8[width * height * 4]);
-	memcpy(baseTexture.get(), image.data.get(), width * height * 4);
+	memcpy(baseTexture.get(), image.data.get(), (uint64)width * height * 4);
 
 	glTexChanged = true;
 }
